requirements.txt
alembic==1.16.5
asttokens==3.0.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dicom2jpg==0.1.10
executing==2.2.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grad-cam==1.5.5
greenlet==3.2.4
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
optuna==4.5.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pydicom==3.0.1
Pygments==2.19.2
pylibjpeg==2.1.0
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-gradcam==0.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
ttach==0.0.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
widgetsnbextension==4.0.14
